# Micro performance optimisation in map matching